remove duplicate UnmaskedAmount definition, update includes
