+++ transaction/core/src/ring_ct/rct_bulletproofs.rs
@@ -32,7 +32,7 @@ use crate::{
     range_proofs::{check_range_proofs, generate_range_proofs},
     ring_ct::{compute_mlsag_signing_digest, Error, ExtendedMessageDigest, GeneratorCache},
     tx::TxPrefix,
-    Amount, BlockVersion, TxSummary,
+    Amount, BlockVersion, TxSummary, UnmaskedAmount,
 };
 
 /// A presigned RingMLSAG and ancillary data needed to incorporate it into a
@@ -77,6 +77,25 @@ pub struct OutputSecret {
     pub amount: Amount,
     /// The blinding factor of the output we are creating
     pub blinding: Scalar,
+}
+
+impl From<UnmaskedAmount> for OutputSecret {
+    fn from(src: UnmaskedAmount) -> Self {
+        Self {
+            amount: Amount::new(src.value, src.token_id.into()),
+            blinding: src.blinding.into(),
+        }
+    }
+}
+
+impl From<OutputSecret> for UnmaskedAmount {
+    fn from(src: OutputSecret) -> Self {
+        Self {
+            value: src.amount.value,
+            token_id: *src.amount.token_id,
+            blinding: src.blinding.into(),
+        }
+    }
 }
 
 /// The parts of a TxIn needed to validate a corresponding MLSAG
