+++ transaction/extra/src/unsigned_tx.rs
@@ -1,6 +1,6 @@
 // Copyright (c) 2018-2022 The MobileCoin Foundation
 
-use crate::{TxOutSummaryUnblindingData, TxSummaryUnblindingData, UnmaskedAmount};
+use crate::{TxOutSummaryUnblindingData, TxSummaryUnblindingData};
 use alloc::vec::Vec;
 use mc_crypto_ring_signature_signer::RingSigner;
 use mc_transaction_core::{
@@ -11,7 +11,9 @@ use mc_transaction_core::{
     tx::{Tx, TxPrefix},
     FeeMap,
 };
-use mc_transaction_types::{amount::Amount, tx_summary::TxSummary, BlockVersion, TokenId};
+use mc_transaction_types::{
+    amount::Amount, tx_summary::TxSummary, unmasked_amount::UnmaskedAmount, BlockVersion, TokenId,
+};
 use rand_core::{CryptoRng, RngCore};
 use serde::{Deserialize, Serialize};
